Get rid of heaps of warnings. Use QSizeF instead of QSize
svn path=/trunk/extragear/multimedia/amarok/; revision=679605

+++ src/contextview/contextview.cpp
@@ -247,7 +247,7 @@ void ContextView::resizeEvent( QResizeEvent *event )
     {
         ContextBox *box = dynamic_cast<ContextBox*>( item );
         if( box )
-            box->ensureWidthFits( newSize.width() );
+            box->ensureWidthFits( (qreal)newSize.width() );
     }
 }
